rearranging and renaming TurmaDTO
